requirements.txt
flask==2.3.2
twilio==9.0.0
groq>=0.3.0
openai-whisper
pyttsx3==2.90
python-dotenv==1.0.0
pandas==2.0.3
googletrans==4.0.0-rc1
google-cloud-translate==2.0.1
flask-cors==3.0.10
requests>=2.31.0
